+++ R/read-png.R
@@ -36,3 +36,11 @@ read_png_nara <- function(raw_vec, flags = 0) {
 read_png_raster <- function(raw_vec, flags = 0) {
   .Call(read_png_raster_, raw_vec, as.integer(flags))
 }
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' @rdname read_png_raw
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+read_png_rgba <- function(raw_vec, flags = 0) {
+  .Call(read_png_rgba_, raw_vec, as.integer(flags))
+}