+++ src/init.c
@@ -6,6 +6,7 @@
 extern SEXP read_png_raw_   (SEXP raw_vec_, SEXP fmt_, SEXP flags_);
 extern SEXP read_png_nara_  (SEXP raw_vec_, SEXP flags_);
 extern SEXP read_png_raster_(SEXP raw_vec_, SEXP flags_);
+extern SEXP read_png_rgba_  (SEXP raw_vec_, SEXP flags_);
 
 extern SEXP extract_png_info_(SEXP raw_vec_);
 
@@ -14,6 +15,7 @@ static const R_CallMethodDef CEntries[] = {
   {"read_png_raw_"   , (DL_FUNC) &read_png_raw_   , 3},
   {"read_png_nara_"  , (DL_FUNC) &read_png_nara_  , 2},
   {"read_png_raster_", (DL_FUNC) &read_png_raster_, 2},
+  {"read_png_rgba_"  , (DL_FUNC) &read_png_rgba_  , 2},