+++ src/read-png.c
@@ -195,7 +195,7 @@ SEXP read_png_raster_(SEXP raw_vec_, SEXP flags_) {
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // 
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~x`~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   SEXP dims_ = PROTECT(allocVector(INTSXP, 2));
   INTEGER(dims_)[0] = ihdr.height;
   INTEGER(dims_)[1] = ihdr.width;
@@ -211,15 +211,97 @@ SEXP read_png_raster_(SEXP raw_vec_, SEXP flags_) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Read PNG as RGBA array
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP read_png_rgb_(SEXP raw_vec_, SEXP flags_) {
-  return R_NilValue;
-}
-
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Read PNG as RGBA array
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP read_png_grey_(SEXP raw_vec_, SEXP flags_) {
-  return R_NilValue;
+SEXP read_png_rgba_(SEXP raw_vec_, SEXP flags_) {
+  size_t out_size;
+  int buf_size = length(raw_vec_);
+  
+  unsigned char *buf = (unsigned char *)RAW(raw_vec_);
+  int fmt   = SPNG_FMT_RGBA8;
+  int flags = INTEGER(flags_)[0];
+  
+  
+  /* Create a context */
+  spng_ctx *ctx = spng_ctx_new(0);
+  
+  
+  /* Set memory usage limits for storing standard and unknown chunks,
+   this is important when reading arbitrary files! */
+  size_t limit = 1024 * 1024 * 64;
+  spng_set_chunk_limits(ctx, limit, limit);
+  
+  /* Set an input buffer */
+  spng_set_png_buffer(ctx, buf, buf_size);
+  
+  /* Determine output image size */
+  spng_decoded_image_size(ctx, fmt, &out_size);
+  
+  
+  unsigned char * out = (unsigned char *)malloc(out_size);
+  if (out == NULL) {
+    error("Couldn't allocate PNG buffer");
+  }
+  
+  // get info
+  struct spng_ihdr ihdr;
+  int err = spng_get_ihdr(ctx, &ihdr);
+  if (err) {
+    spng_ctx_free(ctx);
+    error("spng_get_ihdr() error: %s\n", spng_strerror(err));
+  }
+  
+  /* Decode to 8-bit RGBA */
+  err = spng_decode_image(ctx, out, out_size, fmt, flags);
+  if (err) {
+    spng_ctx_free(ctx);
+    error("spng_decode_image() error: %s\n", spng_strerror(err));
+  }
+  
+  /* Free context memory */
+  spng_ctx_free(ctx);
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Format raw bytes as 3D array:  width, height, depth
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  SEXP out_ = PROTECT(allocVector(REALSXP, out_size));
+  
+  double *ptr = REAL(out_);
+  unsigned char *raw = out;
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Need to switch from raw data (row-major) to R array (column-major)
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  for (int row = 0; row < ihdr.height; row++) {
+    
+    double *r   = ptr + row;
+    double *g   = ptr + row + (out_size >> 2) * 1;
+    double *b   = ptr + row + (out_size >> 2) * 2;
+    double *a   = ptr + row + (out_size >> 2) * 3;
+    
+    for (int col = 0; col < ihdr.width; col++) {
+      *r = *raw++ / 255.0;
+      *g = *raw++ / 255.0;
+      *b = *raw++ / 255.0;
+      *a = *raw++ / 255.0;
+      r += ihdr.height;
+      g += ihdr.height;
+      b += ihdr.height;
+      a += ihdr.height;
+    }
+  }
+  
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // 
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~x`~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  SEXP dims_ = PROTECT(allocVector(INTSXP, 3));
+  INTEGER(dims_)[0] = ihdr.height;
+  INTEGER(dims_)[1] = ihdr.width;
+  INTEGER(dims_)[2] = 4;
+  
+  setAttrib(out_, R_DimSymbol, dims_);
+  setAttrib(out_, R_ClassSymbol, mkString("array"));
+  
+  UNPROTECT(2);
+  return out_;
 }